4th_semester/ISP/3rd-4th_lr/project/requirements.txt
Django
psycopg2
Pillow
PyYaml
django-crispy-forms
pytest-django
Coverage
pytest
mixer
pytest-cov
pytest-xdist
djangorestframework
gunicorn